+++ src/components/general-widgets/SearchBar/utils/getSearchBarModal.spec.js
@@ -1,15 +1,5 @@
 import React from 'react';
-import { shallow } from 'enzyme';
-import {
-  expectComponentToHaveChildren,
-  expectComponentToHaveProps,
-} from '@lodgify/enzyme-jest-expect-helpers';
-import { Modal as SemanticModal, Form } from 'semantic-ui-react';
-import { expectComponentToBe } from '@lodgify/enzyme-jest-expect-helpers/lib/expectComponentToBe';
-
-import { Heading } from 'typography/Heading';
-import { Modal } from 'elements/Modal';
-import { Divider } from 'elements/Divider';
+import { mount } from 'enzyme';
 
 import { getSearchBarModal } from './getSearchBarModal';
 
@@ -24,7 +14,7 @@ const handleSubmit = () => {};
 const persistInputChange = () => {};
 
 const getModalMarkup = extraProps =>
-  shallow(
+  mount(
     <div>
       {getSearchBarModal(
         {
@@ -35,87 +25,22 @@ const getModalMarkup = extraProps =>
         persistInputChange
       )}
     </div>
-  ).children();
+  );
 
 describe('getSearchBarModal', () => {
-  it('should return a `Modal`', () => {
-    const wrapper = getModalMarkup();
+  it('should return the right structure', () => {
+    const actual = getModalMarkup();
 
-    expectComponentToBe(wrapper, Modal);
-  });
-
-  describe('the `Modal`', () => {
-    it('should have the right props', () => {
-      const wrapper = getModalMarkup();
-      const { isModalOpen, onCloseModal, searchButton } = props;
-
-      expectComponentToHaveProps(wrapper, {
-        isFullscreen: true,
-        isOpen: isModalOpen,
-        onClose: onCloseModal,
-        trigger: searchButton,
-      });
-    });
-
-    it('should have the right children', () => {
-      const wrapper = getModalMarkup();
-
-      expectComponentToHaveChildren(wrapper, SemanticModal.Content);
-    });
-  });
-
-  describe('the `Modal.Content`', () => {
-    it('should have the right children', () => {
-      const wrapper = getModalMarkup()
-        .find(SemanticModal.Content)
-        .at(0);
-
-      expectComponentToHaveChildren(wrapper, Heading, Form);
-    });
-
-    it('should have the right children if mobile summary is defined', () => {
-      const wrapper = getModalMarkup({
-        modalSummaryElement: <div />,
-      })
-        .find(SemanticModal.Content)
-        .at(0);
-
-      expectComponentToHaveChildren(wrapper, 'div', Form);
-    });
-  });
-
-  describe('the first `Heading`', () => {
-    const getHeading = () =>
-      getModalMarkup()
-        .find(Heading)
-        .at(0);
-
-    it('should have the right props', () => {
-      const wrapper = getHeading();
-
-      expectComponentToHaveProps(wrapper, {
-        size: 'small',
-      });
-    });
-
-    it('should have the right children', () => {
-      const wrapper = getHeading();
-
-      expectComponentToHaveChildren(wrapper, props.modalHeadingText);
-    });
+    expect(actual).toMatchSnapshot();
   });
 
   describe('if `props.modalSummaryElement` is passed', () => {
-    describe('the first `div` element', () => {
-      it('should have the right children', () => {
-        const wrapper = getModalMarkup({
-          modalSummaryElement: <div />,
-        })
-          .find('div')
-          .at(0);
-
-        expectComponentToHaveChildren(wrapper, 'div', Divider);
+    it('should return the right structure', () => {
+      const actual = getModalMarkup({
+        modalSummaryElement: <div />,
       });
+
+      expect(actual).toMatchSnapshot();
     });
   });
 });