+++ src/utils/get-gallery-heading-markup/getGalleryHeadingMarkup.js
@@ -1,4 +1,4 @@
-import React, { Fragment } from 'react';
+import React from 'react';
 
 import { Heading } from 'typography/Heading';
 import { Rating } from 'elements/Rating';
@@ -10,8 +10,8 @@ import { Rating } from 'elements/Rating';
  */
 export const getGalleryHeadingMarkup = (propertyName, ratingNumber) =>
   propertyName && (
-    <Fragment>
+    <div className="heading-with-rating">
       <Heading>{propertyName}</Heading>
       {ratingNumber > 0 && <Rating ratingNumber={ratingNumber} />}
-    </Fragment>
+    </div>
   );